+++ src/secstructartist/__init__.py
@@ -1,3 +1,2 @@
 from .constants import SSA_VERSION as __version__
-from .core.secstructartist import SecStructArtist
-from .core.elementartist import ElementArtist
+from .secstructartist import SecStructArtist
