+++ src/secstructartist/drawstyle.py
