+++ src/secstructartist/utils.py
